requirements.txt
bitsandbytes==0.44.1
fire==0.7.0
hyperparse==0.0.6
ipykernel==6.29.5
jaraco.collections==5.1.0
matplotlib==3.7.5
protobuf>=3.20.3,<5 
scikit-learn==1.3.2
sentencepiece==0.2.0
tomli==2.0.1
diffusers==0.30.3
transformers==4.45.2
accelerate==1.0.1
deepface==0.0.93
tensorflow==2.16.1
tf-keras==2.16
torchmetrics[multimodal]==1.6.2
pyiqa==0.1.10
opencv-python
deepface
torchmetrics